refactor: move error-handler to services/exposed
