+++ src/index.js
@@ -20,7 +20,7 @@ const SchemaSerializer = require('./serializers/schema');
 const logger = require('./services/logger');
 const pathService = require('./services/path');
 const Integrator = require('./integrations');
-const errorHandler = require('./services/error-handler');
+const errorHandler = require('./services/exposed/error-handler');
 const ApimapSender = require('./services/apimap-sender');
 const ipWhitelist = require('./services/ip-whitelist');
 const SchemaFileUpdater = require('./services/schema-file-updater');
